requirements.txt
python-telegram-bot[job-queue]>=21.7
feedparser==6.0.11
python-dotenv>=1.0.0
